requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Snowflake connector
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.0

# LLM integration
openai==1.3.5
langchain==0.0.340
tiktoken==0.5.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Security & utilities
python-dotenv==1.0.0
httpx==0.25.1
aiofiles==23.2.1
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1